requirements.txt
langchain
langchain-groq
langchain-core
streamlit
python-dotenv